requirements.txt
ultralytics~=8.2.56
opencv-python-headless
opencv-python~=4.10.0.84
pyyaml~=6.0.1
numpy~=1.26.4
pandas~=2.2.2
matplotlib~=3.9.1
seaborn~=0.13.2
pillow~=10.4.0
torch~=2.3.1